feat(vim): resize terminal when running rust-analyzer via coc
I use nvim-coc and rust-analyzer a lot those days but when running a
`CocCommand` that opens a terminal its size is very small.

My desire is to replace `terminal` with `tmux` because I think it is
more comfortable for my workflow but I don't think it is support just
now.

In the meantime I found a workaround that resize the terminal as soon as
coc opens it